lab tests: Remove parallel execution from all tests
With the move to GitLab things in the tests have sped up and exposed
races.  I'm sure a lot of these are due to the parallel executions of
tests.  Until we have time in the future to examine these closer I am just
going to drop parallel execution of testing.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>